+++ exchange/src/rabbitmq/client_manager/RabbitMQClientManager.hpp
@@ -14,7 +14,7 @@ public:
      * This ensures that all clients are connected to the exchange and have successfully
      * started (vs a RMQ or firebase error)
      */
-    static void waitForClients(manager::ClientManager& manager, int num_clients);
+    static void waitForClients(manager::ClientManager& manager, size_t num_clients);
 
     static void sendStartTime(const manager::ClientManager& manager, int wait_seconds);
 };
